\name{FM2.param}
\alias{FM2.param}

\title{Parameter object}

\description{ 
  Create a parameter object for the simulation data of backcross , F2 or RIL intercross.
} 


\usage{ FM2.param(par_obj, curve_type, cross_type=CROSS_BC, covar_type=COVAR_AR1) }

\arguments{

\item{curve_type}{The computational model, three kinds of curve can be specified. \cr
CURVE_LC: Logistical growth curve. \cr
CURVE_BI: Bi-exponential curve. \cr
CURVE_PC: pharmacology curve.\cr
CURVE_NP: Nonparamtric curve.\cr
CURVE_EXP: Exponentiation curve.\cr
}


\item{par_obj}{The simulation parameter object specified in this package. \cr
par_LC: Parameters for CURVE_LC. \cr
par_BI: Parameters for CURVE_BI. \cr
par_PC: Parameters for CURVE_PC.\cr
par_NP: Parameters for CURVE_NP.\cr
par_EXP: Parameters for CURVE_EXP.\cr
}

\item{cross_type}{Indicate which cross to simulate, two values are valid.\cr
CROSS_BC : backcross.\cr
CROSS_F2: F2 intercross.\cr
CROSS_RIL: RIL by selfing.}

\item{covar_type}{Indicate which cross to simulate, two values are valid.\cr
COVAR_AR1 : AR1 covariance.\cr
COVAR_SAD2: SAD covariance.\cr}

}

\value{
A parameter object of class FM2.par. The structure is described in the details section.
You can use summary(\code{\link{summary.FM2.par}}) command to show these information or customize it in R.
}

\details{

Three models(or curves) are implemented in current version, including:

\bold{1) Logistic Curve(CURVE_LC)}

        \deqn{g(t) = \frac{a}{1+b*e^{-r*t}} }{g(t)=a/(1+b*exp(-rt))}
        
\bold{2) Bi-Exponential Curve(CURVE_BI)}

        \deqn{g(t) = { P_{1}*e^{-d_{1}*t} + P_{2}*e^{-d_{2}*t} } }{g(t)=P1*exp(-d1*t) + P2*exp(-d2*t)}

\bold{3) Pharmacology Curve(CURVE_PC)}

        \deqn{g(t) = \frac{E_{max}*t}{Ec_{50}+t} + E_{0} }{g(t)=Emax*t/(Ec50+t) + E0}

\bold{4) Nonparametric method(CURVE_NP)}

        \deqn{g(t)=u_{0}+u_{1}*t + u_{2}*(2*t^2+1)/2 + u_{3}*(5*t^3-3t)/2...}


\bold{5) Exponentiation curve(CURVE_EXP)}

        \deqn{g(t)= {a*t^b}}{g(t)=a*t^b}






Each model(or curve) has different equation and parameters. 
The following introduces the parameters of every curve which has been implemeneted in this package. 











\bold{1) Logistical growth curve.}

The following report is generated by the summary(\code{\link{summary.FM2.par}}) command. It shows the members in the object for logistic curve. 

\preformatted{
The parameter for FunMap model:
------------------------------------
           Date: 2011-01-18 00:34:20
          Model: Logistic Curve
          Cross: BC
    Sample size: 100       
   Sample times: 7         
    Marker pos.: 0,20,40,60,80,100
       QTL pos.: 50        
       LC(Qq) a: 19.98240  
              b: 8.77680   
              r: 0.46990   
       LC(qq) a: 15.95070  
              b: 7.57370   
              r: 0.48360   
            rho: 0.75430   
         sigma2: 0.22600   
------------------------------------
}

Below is the structure of this parameter object. The members except cross_type and model_type can be customized before the parameter object is used to generate raw data.

\preformatted{
List of 11
 $ name        : chr "LC.BC.par"
 $ cross       : num 1
 $ model       : num 1
 $ simu_N      : num 100
 $ simu_mrkdist: num [1:6] 0 20 20 20 20 20
 $ simu_times  : int [1:7] 1 2 3 4 5 6 7
 $ simu_qtlpos : num 50
 $ simu_rho    : num 0.754
 $ simu_s2     : num 0.226
 $ Qq1         :List of 3
  ..$ simu_a: num 20
  ..$ simu_b: num 8.78
  ..$ simu_r: num 0.47
 $ qq0         :List of 3
  ..$ simu_a: num 16
  ..$ simu_b: num 7.57
  ..$ simu_r: num 0.484
 - attr(*, "class")= chr "FM2.par"
}

\bold{2) Bi-exponential curve.}

The following report is generated by the summary(\code{\link{summary.FM2.par}}) command. It shows the members in the object for Bi-exponential curve. 

\preformatted{
The parameter for FunMap model:
------------------------------------
           Date: 2011-01-18 00:34:50
          Model: Bi-exponential Curve
          Cross: BC
    Sample size: 100       
   Sample times: 10        
    Marker pos.: 0,20,40,60,80,100
       QTL pos.: 50        
       BI(Qq)p1: 8.98240   
             d1: 0.46990   
             p2: 7.77680   
             d2: 1.46990   
       BI(qq)p1: 15.95070  
             d1: 0.68360   
             p2: 10.57370  
             d2: 1.68360   
            rho: 0.75430   
         sigma2: 0.22600   
------------------------------------

}

Below is the structure of this parameter object. The parameters except cross_type and model_type can be customized before the parameter object is used to generate raw data.

\preformatted{
List of 17
 $ name        : chr "BI.BC.par"
 $ cross       : num 1
 $ model       : num 2
 $ simu_Qq1_p1 : num 8.98
 $ simu_Qq1_d1 : num 0.47
 $ simu_Qq1_p2 : num 7.78
 $ simu_Qq1_d2 : num 1.47
 $ simu_qq0_p1 : num 16
 $ simu_qq0_d1 : num 0.684
 $ simu_qq0_p2 : num 10.6
 $ simu_qq0_d2 : num 1.68
 $ simu_rho    : num 0.754
 $ simu_s2     : num 0.226
 $ simu_N      : num 100
 $ simu_mrkdist: num [1:6] 0 20 20 20 20 20
 $ simu_times  : int [1:10] 1 2 3 4 5 6 7 8 9 10
 $ simu_qtlpos : num 50
 - attr(*, "class")= chr "FM2.par"

}

\bold{3) Pharmacology curve.}

The following report is generated by the summary(\code{\link{summary.FM2.par}}) command. It shows the members in the object for Pharmacology curve. 

\preformatted{
The parameter for FunMap model:
------------------------------------
           Date: 2011-01-18 00:35:22
          Model: Pharmacology Curve
          Cross: BC
    Sample size: 100       
   Sample times: 7         
    Marker pos.: 0,20,40,60,80,100
       QTL pos.: 50        
     PC(Qq)  E0: 8.98240   
            E50: 14.46990  
           EMax: 20.77680  
     PC(qq)  E0: 7.95070   
            E50: 14.68360  
           Emax: 20.57370  
            rho: 0.75430   
         sigma2: 0.22600   
------------------------------------

}

Below is the structure of this parameter object. The members except cross_type and model_type can be customized before the parameter object is used to generate raw data.

\preformatted{
List of 15
 $ name         : chr "PC.BC.par"
 $ cross        : num 1
 $ model        : num 3
 $ simu_Qq1_E0  : num 8.98
 $ simu_Qq1_E50 : num 14.5
 $ simu_Qq1_Emax: num 20.8
 $ simu_qq0_E0  : num 7.95
 $ simu_qq0_E50 : num 14.7
 $ simu_qq0_Emax: num 20.6
 $ simu_rho     : num 0.754
 $ simu_s2      : num 0.226
 $ simu_N       : num 100
 $ simu_mrkdist : num [1:6] 0 20 20 20 20 20
 $ simu_times   : int [1:7] 1 2 3 4 5 6 7
 $ simu_qtlpos  : num 50
 - attr(*, "class")= chr "FM2.par"
}

\bold{4) Nonparametric method.}

The following report is generated by the summary(\code{\link{summary.FM2.par}}) command. It shows the members in the object for nonparametric curve. 

\preformatted{
The parameter for FunMap model:
------------------------------------
           Date: 2011-01-18 00:32:26
          Model: Nonparametric method
          Cross: BC
    Sample size: 100       
   Sample times: 9         
    Marker pos.: 0,20,40,60,80,100
       QTL pos.: 50        
       LP Order: 6.00000   
     NP(Qq)  u0: 7.14800   
             u1: 1.37900   
             u2: -4.48900  
             u3: 2.00400   
             u4: 0.66200   
             u5: -0.83600  
             u6: 0.43200   
     NP(qq)  u0: 9.04900   
             u1: 1.15100   
             u2: -6.01900  
             u3: 2.65100   
             u4: 0.65200   
             u5: -0.79700  
             u6: 0.62100   
            rho: 0.75430   
         sigma2: 0.30000   
------------------------------------
}

Below is the structure of this parameter object. The members except cross_type and model_type can be customized before the parameter object is used to generate raw data.

\preformatted{
List of 12
 $ name         : chr "NP.BC.par"
 $ cross        : num 1
 $ model        : num 4
 $ simu_np_order: num 6
 $ Qq1          :List of 7
  ..$ simu_u0: num 7.15
  ..$ simu_u1: num 1.38
  ..$ simu_u2: num -4.49
  ..$ simu_u3: num 2
  ..$ simu_u4: num 0.662
  ..$ simu_u5: num -0.836
  ..$ simu_u6: num 0.432
 $ qq0          :List of 7
  ..$ simu_u0: num 9.05
  ..$ simu_u1: num 1.15
  ..$ simu_u2: num -6.02
  ..$ simu_u3: num 2.65
  ..$ simu_u4: num 0.652
  ..$ simu_u5: num -0.797
  ..$ simu_u6: num 0.621
 $ simu_rho     : num 0.754
 $ simu_s2      : num 0.3
 $ simu_N       : num 100
 $ simu_mrkdist : num [1:6] 0 20 20 20 20 20
 $ simu_times   : int [1:9] 1 2 3 4 5 6 7 8 9
 $ simu_qtlpos  : num 50
 - attr(*, "class")= chr "FM2.par"
}

 
}

\author{Zhong Wang, \email{zzw2@psu.edu} }

\examples{

par <- FM2.param( par_LC, CURVE_LC, CROSS_BC );
#Sample size: 1000
par$simu_N <-1000;
#QTL position: 74
par$simu_qtlpos <-74;
#summary the parameter object
summary(par);

#simulate the data set by the parameter object.
dat <- FM2.simulate(par);
summary( dat );
plot(dat);

}

\keyword{functions}
